requirements.txt
click==7.1.2
dimod==0.9.4
dwave-cloud-client==0.7.4
dwave-hybrid==0.5.0
dwave-inspector==0.2.2
dwave-neal==0.5.4
dwave-networkx==0.8.6
dwave-ocean-sdk==2.4.0
dwave-qbsolv==0.3.0
dwave-system==0.9.6
dwave-tabu==0.3.0
dwavebinarycsp==0.1.2
Flask==1.1.2
homebase==1.0.1
itsdangerous==1.1.0
minorminer==0.1.9
networkx==2.4
numpy==1.19.0
ortools==7.7.7810
penaltymodel==0.16.3
penaltymodel-cache==0.4.1
penaltymodel-lp==0.1.3
penaltymodel-mip==0.2.3
plucky==0.4.3
protobuf==3.12.2
pyqubo==0.4.0
PySocks==1.7.1
scipy==1.5.1
Werkzeug==1.0.1
